New Cardano.Api.IO.Base module to hold File types
